backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
asyncpg==0.30.0
alembic==1.13.3
betterproto==2.0.0b6
cryptography==43.0.0